Fix array results
When experimenting with EnumerateInstances when the result set has
more than 1 result,  all the results are identical.  After tracking
this down it appears to be an issue with returning the pointer for
`instance` in each of the for loops.  In these cases the address is
the same.  Returning the pointer for each of the actual addresses
in the array seems to correct the problem.

Note: This is my first exposure to golang, so this could be totally
wrong, but it's working for me and I only tested EnumerateInstances.
The other corrections were based on the information determined from
finding this one.

Signed-off-by: Tony Asleson <tasleson@redhat.com>